Rewrite backend task implementations to appropriate places
