+++ backend/tests/test_tasks/test_prune_inactive_reservations.py
@@ -9,7 +9,6 @@ from tilavarauspalvelu.enums import AccessType, ReservationStateChoice
 from tilavarauspalvelu.integrations.keyless_entry import PindoraClient
 from tilavarauspalvelu.integrations.keyless_entry.exceptions import PindoraAPIError
 from tilavarauspalvelu.models import Reservation
-from tilavarauspalvelu.services.pruning import prune_inactive_reservations
 from utils.date_utils import local_datetime
 
 from tests.factories import ReservationFactory
@@ -27,7 +26,7 @@ def test_prune_inactive_reservations__deletes_old_reservations_with_state_create
         state=ReservationStateChoice.CREATED,
     )
 
-    prune_inactive_reservations()
+    Reservation.objects.delete_inactive()
 
     assert Reservation.objects.exists() is False
 
@@ -40,7 +39,7 @@ def test_prune_inactive_reservations__does_not_delete_inactive_reservations_with
             continue
         ReservationFactory.create(created_at=twenty_minutes_ago, state=state)
 
-    prune_inactive_reservations()
+    Reservation.objects.delete_inactive()
 
     assert Reservation.objects.count() == len(ReservationStateChoice.choices) - 1
 
@@ -51,7 +50,7 @@ def test_prune_inactive_reservations__does_not_delete_recent_reservations():
         state=ReservationStateChoice.CREATED,
     )
 
-    prune_inactive_reservations()
+    Reservation.objects.delete_inactive()
 
     assert Reservation.objects.exists() is True
 
@@ -65,7 +64,7 @@ def test_prune_inactive_reservations__delete_from_pindora():
         access_code_generated_at=local_datetime(),
     )
 
-    prune_inactive_reservations()
+    Reservation.objects.delete_inactive()
 
     assert Reservation.objects.exists() is False
 
@@ -81,10 +80,10 @@ def test_prune_inactive_reservations__delete_from_pindora__call_fails_runs_task(
         access_code_generated_at=local_datetime(),
     )
 
-    path = "tilavarauspalvelu.services.pruning.delete_pindora_reservation.delay"
+    path = "tilavarauspalvelu.models.reservation.queryset.delete_pindora_reservation.delay"
 
     with mock.patch(path) as task:
-        prune_inactive_reservations()
+        Reservation.objects.delete_inactive()
 
     assert Reservation.objects.exists() is False
