+++ backend/tests/test_tasks/test_prune_recurring_reservations.py
@@ -6,7 +6,6 @@ import pytest
 from freezegun import freeze_time
 
 from tilavarauspalvelu.models import RecurringReservation
-from tilavarauspalvelu.services.pruning import prune_recurring_reservations
 from utils.date_utils import local_datetime
 
 from tests.factories import RecurringReservationFactory
@@ -21,7 +20,7 @@ def test_prune_recurring_reservations__recurring_reservations_without_reservatio
     with freeze_time(local_datetime() - datetime.timedelta(days=1)):
         RecurringReservationFactory.create()
 
-    prune_recurring_reservations()
+    RecurringReservation.objects.delete_empty_series()
 
     assert RecurringReservation.objects.exists() is False
 
@@ -30,7 +29,7 @@ def test_prune_recurring_reservations__recurring_reservations_with_reservations_
     with freeze_time(local_datetime() - datetime.timedelta(days=1)):
         RecurringReservationFactory.create(reservations__name="foo")
 
-    prune_recurring_reservations()
+    RecurringReservation.objects.delete_empty_series()
 
     assert RecurringReservation.objects.exists() is True
 
@@ -39,6 +38,6 @@ def test_prune_recurring_reservations__recurring_reservations_deletion_respects_
     with freeze_time(local_datetime() - datetime.timedelta(hours=23, minutes=59)):
         RecurringReservationFactory.create()
 
-    prune_recurring_reservations()
+    RecurringReservation.objects.delete_empty_series()
 
     assert RecurringReservation.objects.exists() is True
