+++ backend/tests/test_tasks/test_prune_reservation_statistics.py
@@ -4,7 +4,6 @@ import pytest
 from dateutil.relativedelta import relativedelta
 
 from tilavarauspalvelu.models import ReservationStatistic
-from tilavarauspalvelu.services.pruning import prune_reservation_statistics
 from utils.date_utils import local_datetime
 
 from tests.factories import ReservationFactory
@@ -24,7 +23,7 @@ def test_prune_reservation_statistics__deletes_in_the_given_time(settings):
     delete = ReservationFactory.create(created_at=five_years_ago, name="DELETE ME")
     keep = ReservationFactory.create(created_at=four_years_ago, name="KEEP ME")
 
-    prune_reservation_statistics()
+    ReservationStatistic.objects.delete_expired_statistics()
 
     assert ReservationStatistic.objects.filter(reservation=delete).exists() is False
     assert ReservationStatistic.objects.filter(reservation=keep).count() == 1
