+++ backend/tilavarauspalvelu/models/application/actions.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 import dataclasses
 from typing import TYPE_CHECKING
 
+from tilavarauspalvelu.models import AllocatedTimeSlot, ReservationUnitOption
+
 if TYPE_CHECKING:
     from tilavarauspalvelu.models import Application
 
@@ -21,13 +23,7 @@ class ApplicationActions:
         Remove application allocations, and unlock locked reservation unit options.
         Rejected options stay rejected.
         """
-        from tilavarauspalvelu.models import AllocatedTimeSlot, ReservationUnitOption
-
-        ReservationUnitOption.objects.filter(
-            application_section__application=self.application,
-        ).update(
-            locked=False,
-        )
+        ReservationUnitOption.objects.filter(application_section__application=self.application).update(locked=False)
 
         AllocatedTimeSlot.objects.filter(
             reservation_unit_option__application_section__application=self.application,
