+++ backend/tilavarauspalvelu/models/recurring_reservation/actions.py
@@ -2,12 +2,13 @@ from __future__ import annotations
 
 import dataclasses
 import datetime
-from itertools import chain
-from typing import TYPE_CHECKING, Any, TypedDict
+from typing import TYPE_CHECKING
 
+from tilavarauspalvelu.dataclasses import ReservationSeriesCalculationResults
 from tilavarauspalvelu.enums import AccessType, RejectionReadinessChoice
 from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
 from tilavarauspalvelu.models import AffectingTimeSpan, ApplicationSection, RejectedOccurrence, Reservation
+from tilavarauspalvelu.typing import ReservationPeriod
 from utils.date_utils import DEFAULT_TIMEZONE, combine, get_periods_between
 
 if TYPE_CHECKING:
@@ -15,112 +16,13 @@ if TYPE_CHECKING:
 
     from django.db import models
 
-    from tilavarauspalvelu.enums import (
-        CustomerTypeChoice,
-        ReservationStateChoice,
-        ReservationTypeChoice,
-        ReservationTypeStaffChoice,
-    )
-    from tilavarauspalvelu.models import (
-        AgeGroup,
-        City,
-        RecurringReservation,
-        ReservableTimeSpan,
-        ReservationPurpose,
-        User,
-    )
+    from tilavarauspalvelu.models import RecurringReservation, ReservableTimeSpan
+    from tilavarauspalvelu.typing import ReservationDetails
 
 
 __all__ = [
     "RecurringReservationActions",
-    "ReservationDetails",
-    "ReservationSeriesCalculationResults",
-    "ReservationSeriesCalculationResults",
 ]
-
-
-class ReservationPeriod(TypedDict):
-    begin: datetime.datetime
-    end: datetime.datetime
-
-
-@dataclasses.dataclass
-class ReservationSeriesCalculationResults:
-    non_overlapping: list[ReservationPeriod] = dataclasses.field(default_factory=list)
-    overlapping: list[ReservationPeriod] = dataclasses.field(default_factory=list)
-    not_reservable: list[ReservationPeriod] = dataclasses.field(default_factory=list)
-    invalid_start_interval: list[ReservationPeriod] = dataclasses.field(default_factory=list)
-
-    def as_json(self, periods: list[ReservationPeriod]) -> list[dict[str, Any]]:
-        return [
-            {
-                "begin": period["begin"].isoformat(timespec="seconds"),
-                "end": period["end"].isoformat(timespec="seconds"),
-            }
-            for period in periods
-        ]
-
-    @property
-    def overlapping_json(self) -> list[dict[str, Any]]:
-        return self.as_json(self.overlapping)
-
-    @property
-    def not_reservable_json(self) -> list[dict[str, Any]]:
-        return self.as_json(self.not_reservable)
-
-    @property
-    def invalid_start_interval_json(self) -> list[dict[str, Any]]:
-        return self.as_json(self.invalid_start_interval)
-
-    @property
-    def possible(self) -> Iterable[ReservationPeriod]:
-        return self.non_overlapping
-
-    @property
-    def not_possible(self) -> Iterable[ReservationPeriod]:
-        return chain(self.overlapping, self.not_reservable, self.invalid_start_interval)
-
-
-class ReservationDetails(TypedDict, total=False):
-    name: str
-    description: str
-    num_persons: int
-    state: ReservationStateChoice
-    type: ReservationTypeChoice | ReservationTypeStaffChoice
-    working_memo: str
-
-    buffer_time_before: datetime.timedelta
-    buffer_time_after: datetime.timedelta
-    handled_at: datetime.datetime
-    confirmed_at: datetime.datetime
-
-    applying_for_free_of_charge: bool
-    free_of_charge_reason: bool
-
-    reservee_id: str
-    reservee_first_name: str
-    reservee_last_name: str
-    reservee_email: str
-    reservee_phone: str
-    reservee_organisation_name: str
-    reservee_address_street: str
-    reservee_address_city: str
-    reservee_address_zip: str
-    reservee_is_unregistered_association: bool
-    reservee_type: CustomerTypeChoice
-
-    billing_first_name: str
-    billing_last_name: str
-    billing_email: str
-    billing_phone: str
-    billing_address_street: str
-    billing_address_city: str
-    billing_address_zip: str
-
-    user: int | User
-    purpose: int | ReservationPurpose
-    home_city: int | City
-    age_group: int | AgeGroup
 
 
 @dataclasses.dataclass(slots=True, frozen=True)
