+++ backend/tilavarauspalvelu/models/reservation_unit_pricing/queryset.py
@@ -10,6 +10,8 @@ from utils.date_utils import local_date
 if TYPE_CHECKING:
     import datetime
 
+    from tilavarauspalvelu.models import TaxPercentage
+
 __all__ = [
     "ReservationUnitPricingManager",
     "ReservationUnitPricingQuerySet",
@@ -17,7 +19,7 @@ __all__ = [
 
 
 class ReservationUnitPricingQuerySet(models.QuerySet):
-    def exclude_past(self) -> models.QuerySet:
+    def exclude_past(self) -> Self:
         """Return all currently active and future pricings."""
         today = local_date()
 
@@ -55,6 +57,31 @@ class ReservationUnitPricingQuerySet(models.QuerySet):
                 ),
             )
             .filter(row_number=1)
+        )
+
+    def latest_pricings_for_tax_update(self, change_date: datetime.date, ignored_company_codes: list[str]) -> Self:
+        """Get last pricing for each reservation unit before the change date"""
+        return (
+            self.filter(
+                models.Q(begins__lte=change_date, highest_price=0)  # Ignore FREE pricings after the change date
+                | models.Q(highest_price__gt=0)
+            )
+            .exclude(reservation_unit__payment_accounting__company_code__in=ignored_company_codes)
+            .exclude(
+                # Use Unit's Payment Accounting, only if Reservation Unit's Payment Accounting is not set
+                reservation_unit__unit__payment_accounting__company_code__in=ignored_company_codes,
+                reservation_unit__payment_accounting__isnull=True,
+            )
+            .order_by("reservation_unit_id", "-begins")
+            .distinct("reservation_unit_id")
+        )
+
+    def pricings_with_tax_percentage(self, after_date: datetime.date, tax_percentage: TaxPercentage) -> Self:
+        """Get pricings for the given tax percentage from the given date."""
+        return self.filter(
+            begins__gte=after_date,
+            tax_percentage=tax_percentage,
+            highest_price__gt=0,
         )
