+++ backend/tilavarauspalvelu/models/unit/queryset.py
@@ -4,6 +4,7 @@ from typing import Literal, Self
 
 from django.db import models
 
+from tilavarauspalvelu.models import Reservation, ReservationUnit, UnitGroup
 from utils.db import SubqueryCount
 
 __all__ = [
@@ -14,8 +15,6 @@ __all__ = [
 
 class UnitQuerySet(models.QuerySet):
     def order_by_unit_group_name(self, *, language: Literal["fi", "en", "sv"], desc: bool = False) -> Self:
-        from tilavarauspalvelu.models import UnitGroup
-
         return self.alias(**{
             f"unit_group_name_{language}": models.Subquery(
                 queryset=(
@@ -28,8 +27,6 @@ class UnitQuerySet(models.QuerySet):
         }).order_by(models.OrderBy(models.F(f"unit_group_name_{language}"), descending=desc))
 
     def order_by_reservation_units_count(self, *, desc: bool = False) -> Self:
-        from tilavarauspalvelu.models import ReservationUnit
-
         return self.alias(
             reservation_units_count=SubqueryCount(
                 ReservationUnit.objects.filter(unit=models.OuterRef("pk"), is_archived=False).values("id"),
@@ -37,8 +34,6 @@ class UnitQuerySet(models.QuerySet):
         ).order_by(models.OrderBy(models.F("reservation_units_count"), descending=desc))
 
     def order_by_reservation_count(self, *, desc: bool = False) -> Self:
-        from tilavarauspalvelu.models import Reservation
-
         return self.alias(
             reservation_count=SubqueryCount(
                 Reservation.objects.filter(reservation_units__unit=models.OuterRef("pk")).values("id"),
